Move to Shared - SqlSer.cs (#1313)
